% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grp_55.R
\name{age_grp_55}
\alias{age_grp_55}
\title{Generate age groups.}
\usage{
age_grp_55(x)
}
\arguments{
\item{x}{age variable}
}
\description{
Generate age groups based on 'age' variable, up to 55+. Commonly used in HC reports.
}
