% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_adls_token.R
\name{az_adls_token}
\alias{az_adls_token}
\title{Azure authentication}
\usage{
az_adls_token(mode = "silent")
}
\arguments{
\item{mode}{silent, token, or debug. "silent" is the default and is used inside other functions. "token" will return the actual token. "debug" returns useful information for troubleshooting token issues.}
}
\description{
Generate a manual, non-refreshable Azure authentication token with user's account credentials. It expires after an hour and a half. The first time a user runs this function they'll be prompted to sign in via a pop-up window. Afterward, it will happen automatically. This requires Azure CLI to be installed and findable.
}
