% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_locate.R
\name{az_locate}
\alias{az_locate}
\title{Azure authentication}
\usage{
az_locate(path_manual = NULL)
}
\arguments{
\item{path_manual}{default is NULL; if supplied, will check for Azure CLI at that path in addition to the default ones. Not generally necessary.}
}
\description{
This function will ensure Azure CLI can be found via the command line. If not, it will attempt to find and connect. It runs all data lake functions to ensure the necessary authentication is possible.
}
