% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{base_five}
\alias{base_five}
\title{Base Five Color Palette.}
\format{
An object of class \code{character} of length 5.
}
\usage{
base_five
}
\description{
Commonly used default colors
}
\keyword{datasets}
