% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{base_four}
\alias{base_four}
\title{Base Four Color Palette.}
\format{
An object of class \code{character} of length 4.
}
\usage{
base_four
}
\description{
Commonly used default colors
}
\keyword{datasets}
