% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{base_two}
\alias{base_two}
\title{Base Two Color Palette.}
\format{
An object of class \code{character} of length 2.
}
\usage{
base_two
}
\description{
Commonly used default colors
}
\keyword{datasets}
