% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_age_grp.R
\name{census_age_grp}
\alias{census_age_grp}
\title{Generate age groups.}
\usage{
census_age_grp(x)
}
\arguments{
\item{x}{age variable}
}
\description{
Generate age groups based on 'age' variable, matching Census standards.
}
