% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expectations.R
\name{check_expectations}
\alias{check_expectations}
\title{Data validation}
\usage{
check_expectations(data, fail = TRUE, export = TRUE, export_keep = NULL)
}
\arguments{
\item{data}{a dataframe with at least one metadata column created by the "expect_" functions}

\item{fail}{should violations result in an error? Default is TRUE.}

\item{export}{should a dataframe named violations_export of up to 1,000 violations be created? Default is TRUE.}

\item{export_keep}{by default the violations dataframe will keep any metadata columns with violations and the corresponding data. Here, specify any additional columns that should be retained (such as a unique ID). Default is NULL.}
}
\description{
Evaluate the boolean columns created using an expect_ functions, optionally stopped code execution and/or exporting violations to a dataframe for review.
}
