% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_antibiotics.R
\name{clean_antibiotics}
\alias{clean_antibiotics}
\title{Clean STI antibiotic drugs}
\usage{
clean_antibiotics(
  data,
  input = "antibiotic_treatment",
  output = "antibiotic_treatment"
)
}
\arguments{
\item{data}{a dataframe}

\item{input}{the name of the input variable}

\item{output}{the name assigned to the cleaned, output variable}
}
\description{
Clean the names of antibiotic treatment regimens in the chlamydia, gonorrhea, and syphilis data. Some prior data cleaning may be required, such as making sure multiple drugs in the same column are separated by commas, and commas are only used for this purpose. The function can be updated as more drugs are added.
}
