% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_schema.R
\name{clean_schema}
\alias{clean_schema}
\title{Edit Schema of Arrow data}
\usage{
clean_schema(arrow_dataset)
}
\arguments{
\item{arrow_dataset}{data in Arrow environment}
}
\description{
Clean the schema of an arrow dataset to remove "null" types which lead to errors. A typical workflow for this function is:
}
