% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cols.R
\name{compare_cols}
\alias{compare_cols}
\title{Compare column types between DF.}
\usage{
compare_cols(df1, df2)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{A dataframe}
}
\description{
ID differences in column type before combining dataframes. It will also identify columns not present in both dataframes.
}
