% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_display.R
\name{conditional_display}
\alias{conditional_display}
\title{Display dataset}
\usage{
conditional_display(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
Intended for use in Databricks only. Display() returns an error if dataframe is empty. Using this modified version to avoid that to prevent script-stopping errors.
}
