% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_dictionary.R
\name{consult_dictionary}
\alias{consult_dictionary}
\title{Set csv column types}
\usage{
consult_dictionary(
  dictionary,
  column_names,
  column_types,
  date_format = "\%m/\%d/\%Y",
  datetime_format = "\%m/\%d/\%Y \%I:\%M:\%S \%p"
)
}
\arguments{
\item{dictionary}{a dataframe with information from a data dictionary}

\item{column_names}{a column containing expected data columns}

\item{date_format}{a single date format or a vector of multiple date formats used to read in date columns. If a vector is supplied, its length must equal the number of date columns.}

\item{datetime_format}{a single datetime format or a vector of multiple datetime formats used to read in datetime columns. If a vector is supplied, its length must equal the number of datetime columns.}

\item{na_accept}{a column containing expected type for each expected column}
}
\description{
Reference column types specified in a data dictionary to create a column specification (.cols) for read_csv's col_types.
}
