% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_group.R
\name{create_age_group}
\alias{create_age_group}
\title{Generate age groups.}
\usage{
create_age_group(age, group_size = 5, group_min = 10, group_max = 85)
}
\arguments{
\item{age}{age column in a dataframe}

\item{group_size}{the size of the age groups (besides the minimum and maximum groups)}

\item{group_min}{the upper cutoff of the bottom group (i.e 10 for "<10")}

\item{group_max}{the bottom cutoff for the top group (i.e. 85 for "85+" )}
}
\value{
a character vector.
}
\description{
Generate age groups based on age value. Most potential age group schemes should be achievable, barring those with unequally sized buckets. Function defaults match Census standards.
}
\examples{
ages <- data.frame(age = 1:100)
create_age_group(ages[12,])

Or use inside of mutate()
dplyr::mutate(ages, age_group = create_age_group(age))
}
