% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_format.R
\name{expect_format}
\alias{expect_format}
\title{Data validation}
\usage{
expect_format(data, value_pairs, na_accept = FALSE, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{value_pairs}{a named list of columns and their acceptable format (using regex/string comparison)}

\item{na_accept}{should NA values be considered valid?}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check whether columns values are not in the expected format. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
