% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_missing.R
\name{expect_missing}
\alias{expect_missing}
\title{Data validation}
\usage{
expect_missing(data, value_pairs, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{value_pairs}{a list where names are column names and values are the maximum expected missing percentage (between 0 and 1)}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check the percentage of column values that are missing does not exceed a threshold. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
