% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_range.R
\name{expect_range}
\alias{expect_range}
\title{Data validation}
\usage{
expect_range(data, value_pairs, na_accept = FALSE, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{value_pairs}{a named list of columns and their acceptable range (a vector with a minimum and maximum number or date in the YMD format)}

\item{na_accept}{should NA values be considered valid?}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check whether column values are outside of the expected range. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
