% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_type.R
\name{expect_type}
\alias{expect_type}
\title{Data validation}
\usage{
expect_type(data, value_pairs, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{value_pairs}{a list where names are column types (numeric, character, date, logical, or factor) and values are a vector of columns}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check that a column is of the expected type. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
