% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_unique.R
\name{expect_unique}
\alias{expect_unique}
\title{Data validation}
\usage{
expect_unique(data, columns, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{columns}{a vector of columns or of vectors of columns. A nested vector can have up to 4 columns.}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check whether there are duplicate values in a column. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
