% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_values.R
\name{expect_values}
\alias{expect_values}
\title{Data validation}
\usage{
expect_values(data, value_pairs, na_accept = FALSE, warn = FALSE)
}
\arguments{
\item{data}{a dataframe with data to validate}

\item{value_pairs}{a named list of columns and their acceptable values}

\item{na_accept}{should NA values be considered valid?}

\item{warn}{should the function itself warn about violations? Alternatively, use check_expectations().}
}
\description{
Check whether unexpected values are included in a column. This function with create a boolean row that indicates unexpected values, and can be combined with check_expectations to halt a pipeline or export violations for review.
}
