% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_city.R
\name{fix_city}
\alias{fix_city}
\title{Clean city names.}
\usage{
fix_city(data, variable, path)
}
\arguments{
\item{data}{a dataframe}

\item{variable}{name of city variable}

\item{path}{path to CityNameFix.txt}
}
\description{
Cleans city names of common errors using the CityNameFix.txt file.
}
