% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_nicknames.R
\name{flag_nicknames}
\alias{flag_nicknames}
\title{Adjust probabilistic matching.}
\usage{
flag_nicknames(data, name1 = "first_name_1", name2 = "first_name_2")
}
\arguments{
\item{data}{a dataframe}

\item{name1}{name of first matched first name field}

\item{name2}{name of second matched first name field}
}
\description{
Flag matches where first names do not match, but are a valid given name - nickname pair. This flag could then be used to adjust weighting or to allow variable weighting. This is useful because when given names and nicknames are quite different (i.e. Theodore, Ted) matches may not have a very high weight.
}
