% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kv_secret.R
\name{get_kv_secret}
\alias{get_kv_secret}
\title{Get secret.}
\usage{
get_kv_secret(secret_name)
}
\arguments{
\item{secret_name}{the name (key) of a secret}
}
\value{
a string that cannot be directly printed.
}
\description{
This function will get secret values from a Databricks secret scope/key vault. While the base Databricks utility will get the secret from a specified scope, this function will cycle through all scopes to attempt find an accessible version for the user. It's intended to make code more flexible between users with access to different key vaults.
}
