% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{hc_categorical}
\alias{hc_categorical}
\title{HC Categorical Color Palette.}
\format{
An object of class \code{character} of length 7.
}
\usage{
hc_categorical
}
\description{
Commonly used default colors
}
\keyword{datasets}
