% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{hc_diverging}
\alias{hc_diverging}
\title{HC Diverging Color Palette.}
\format{
An object of class \code{character} of length 12.
}
\usage{
hc_diverging
}
\description{
Commonly used default colors
}
\keyword{datasets}
