% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{heat_pal}
\alias{heat_pal}
\title{Default Red Heat Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
heat_pal
}
\description{
Commonly used default colors
}
\keyword{datasets}
