% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_date.R
\name{impute_date}
\alias{impute_date}
\title{Impute missing date information.}
\usage{
impute_date(x, month = "06", day = "15")
}
\arguments{
\item{x}{a dataframe}

\item{month}{the value imputed for a missing month}

\item{year}{the value imputed for a missing day}
}
\description{
Impute missing information for dates in character, YYYYMMDD, format. Used in HIV and HIV D2C scripts. Not useful for dates in other formats.
}
