% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_address.R
\name{is_address}
\alias{is_address}
\title{Identify addresses}
\usage{
is_address(x)
}
\arguments{
\item{x}{a string that may or may not be an address}
}
\description{
Identify whether a string is a valid address using (about) the most permissive criteria: it begins with at least two numbers that are at some point followed by a space and a text character.
}
