% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volumes.R
\name{list_volumes}
\alias{list_volumes}
\title{List Databricks volumes}
\usage{
list_volumes(catalogue = "prd_ph_source", schema = "health_fs", path = NULL)
}
\arguments{
\item{catalogue}{the catalogue that includes the volumes; points to health_fs by default}

\item{schema}{the schema that includes the volumes; points to health_fs by default}

\item{path}{all or part of a specific data lake file path that the user wishes find the corresponding volumes for}
}
\description{
Use SQL queries to list Databricks volumes, their permission scheme, and the corresponding data lake location. Returns a tibble.
}
