% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dates.R
\name{match_dates}
\alias{match_dates}
\title{Match date values}
\usage{
match_dates(data, date1, date2, exact = FALSE, switch = TRUE, threshold = 5)
}
\arguments{
\item{data}{A linked dataframe based on probablistic matching of non-date fields.}

\item{date1}{The date from the first matched dataset}

\item{date2}{The date from the second matched dataset}

\item{exact}{Should only exact matches be retained? Default value is FALSE.}

\item{switch}{Should pairs that match once day and month are reversed in one date be kept? Default is TRUE.}

\item{threshold}{How many days is an acceptable difference between two dates? Default is 5, set to 0 to ignore.}
}
\description{
This function is used to match dates when doing probabalistic matched. Rather than doing simple string or numeric comparison, it includes a few options through which potential matches can be evaluated.
}
