% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_multipart_names.R
\name{match_multipart_names}
\alias{match_multipart_names}
\title{Adjust probabilistic matching.}
\usage{
match_multipart_names(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
Using fields created in clean_match, enhances matching accuracy by better handling potential matches where at least one record has a multi-part last name. If only one last name is multi-part, it tests whether the single name is part of the other. If both are, it tests whether reversing one of the names results in a direct match. It produces the variables \code{last_partial} and \code{last_switched} which can be used when filtering matches to those we accept.
}
