% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_skim.R
\name{mini_skim}
\alias{mini_skim}
\title{Skim dataframe}
\usage{
mini_skim(x, display = TRUE)
}
\arguments{
\item{.data}{a dataframe}

\item{status}{the text to print}
}
\description{
Simple version of skimr::skim so I don't need to load the package.
}
