% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpls_nbhds}
\alias{mpls_nbhds}
\title{Minneapolis neighborhood shapefiles}
\format{
Not going to document everything here.
\describe{
\item{nbhd_name}{Name of neighborhood}
\item{nbhd_number}{Number of neighborhood}
\item{geometry}{Neighborhood geometry object}
...
}
}
\usage{
mpls_nbhds
}
\description{
Shapefiles downloaded from Minneapolis open data. Last updated on 02/24.
}
\keyword{datasets}
