% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{na_color}
\alias{na_color}
\title{Default Missing Color}
\format{
An object of class \code{character} of length 1.
}
\usage{
na_color
}
\description{
Commonly used default colors
}
\keyword{datasets}
