% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_print.R
\name{pipe_print}
\alias{pipe_print}
\title{Print text.}
\usage{
pipe_print(.data, status)
}
\arguments{
\item{.data}{a dataframe}

\item{status}{the text to print}
}
\description{
Print text in the middle of a pipe.
}
