% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmi_leading_zero.R
\name{pmi_leading_zero}
\alias{pmi_leading_zero}
\title{Adjust PMI values}
\usage{
pmi_leading_zero(df, x)
}
\arguments{
\item{df}{A dataframe}

\item{x}{PMI column/value}
}
\description{
Add leadings zeroes that are missing from PMI values.
}
