% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Read Qualtrics data.}
\usage{
read_qualtrics(path, ...)
}
\arguments{
\item{path}{file path}
}
\description{
Read and clean csv files in the standard Qualtrics export format.
}
