% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adls.R
\name{read_raw_adls}
\alias{read_raw_adls}
\alias{read_csv_adls}
\alias{read_parquet_adls}
\alias{read_excel_adls}
\title{Read data lake files}
\usage{
read_raw_adls(
  path,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  return_excel = FALSE
)

read_csv_adls(
  path,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  ...
)

read_parquet_adls(
  path,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  ...
)

read_excel_adls(
  path,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  ...
)
}
\arguments{
\item{path}{the path to a file}

\item{endpoint}{the data lake storage account. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{container}{the data lake container. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{return_excel}{FALSE by default; switch to TRUE in \code{read_adls_raw} if returning raw output from an Excel file.}
}
\description{
\code{read_csv_adls}, \code{read_parquet_adls}, \code{read_excel_adls}, and \code{read_raw_adls} offer functionality to read Azure data lake files. Use ... to pass arguments to the underlying read functions. See SharePoint documentation for more information about how each function works and required setup. These functions are only intended to work with trusted files. Code that transforms raw or refined files should be stored in Databricks.
}
