% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{reference}
\alias{reference}
\title{Data lake folder path.}
\format{
An object of class \code{character} of length 1.
}
\usage{
reference
}
\description{
Trusted reference folder path
}
\keyword{datasets}
