% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hennepin.R
\name{theme_hennepin}
\alias{theme_hennepin}
\title{Hennepin County-inspired ggplot2 theme}
\usage{
theme_hennepin(db_path = NULL)
}
\description{
Sensible default theme based on Hennepin County visualization guide. Can use package color palettes to supplement.
}
