% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_snake_case.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Data cleaning}
\usage{
to_snake_case(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
Convert column names to snake_case. Similar functionality to janitor::clean_names()
}
