% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_adls.R
\name{write_csv_adls}
\alias{write_csv_adls}
\alias{write_parquet_adls}
\title{Write data lake files}
\usage{
write_csv_adls(
  x,
  file,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  ...
)

write_parquet_adls(
  x,
  file,
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  ...
)
}
\arguments{
\item{x}{a dataframe to be written}

\item{endpoint}{the data lake storage account. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{container}{the data lake container. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{path}{the file path}
}
\description{
\code{write_csv_adls} and \code{write_parquet_adls} write  Azure data lake files. Use ... to pass arguments to the underlying write functions. See SharePoint documentation for more information about how each function works and required setup. These functions are only intended to work with trusted files. Code that transforms raw or refined files should be stored in Databricks.
}
