% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_city.R
\name{fix_city}
\alias{fix_city}
\title{Clean city names.}
\usage{
fix_city(data, x)
}
\arguments{
\item{data}{a dataframe}

\item{x}{name of city variable}
}
\description{
Cleans all city names of common errors using the CityNameFix.txt file.
}
