% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{hc_sequential}
\alias{hc_sequential}
\title{Data lake volume path.}
\format{
An object of class \code{character} of length 10.
}
\usage{
hc_sequential
}
\description{
Beginning of all volume paths
}
\keyword{datasets}
