% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_adls.R
\name{list_files_adls}
\alias{list_files_adls}
\title{List data lake files}
\usage{
list_files_adls(
  path = "",
  endpoint = Sys.getenv("ADLS_ENDPOINT"),
  container = Sys.getenv("ADLS_CONTAINER"),
  recursive = FALSE,
  include_dirs = FALSE
)
}
\arguments{
\item{path}{a directory path}

\item{endpoint}{the data lake storage account. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{container}{the data lake container. Pulls from R environmental variable by default. See setup documentation with information about how to set.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{include_dirs}{logical. Should subdirectory names be included?}
}
\description{
List files in an Azure data lake. See SharePoint documentation for more information about how this function works and required setup. This function is intended to only work with trusted files.
}
