% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volumes.R
\name{list_volumes}
\alias{list_volumes}
\title{List Databricks volumes}
\usage{
list_volumes(catalog, schema, parent_folder = NULL)
}
\arguments{
\item{catalog}{the catalog that includes the volumes}

\item{schema}{the schema that includes the volumes}

\item{parent_folder}{optional filter; the parent folder the volume maps to}
}
\description{
Use SQL queries to list Databricks volumes and their permission scheme in a tibble. Use to identify the volume that contains particular data or create a vector of volumes that could be looped through.
}
